Add/update comparison methods for Atom, Bond, Molecule, Species
Python 3 no longer allows implicit sorting of custom classes.

This implements hash, eq, lt, gt special methods for these classes.

In general, the goal was to maintain prior behavior for hash and eq
and implement reasonable behavior for gt and lt.

Sorting of these classes is generally done to enable deterministic behavior,
rather than any quantitative reason, so that was the main consideration.